+++ src/napari_stress/_approximation/__init__.py
@@ -1,14 +1,17 @@
 """Approximation subpackage for napari-stress."""
 
-from .expansion import (
-    EllipsoidExpander,
-    EllipsoidImageExpander,
+from .expansion_spherical_harmonics import (
     SphericalHarmonicsExpander,
+)
+
+from .expansion_ellipsoid import (
+    EllipsoidImageExpander,
 )
 from .expansion_napari import (
     expand_ellipsoid_on_image,
     expand_points_on_fitted_ellipsoid,
     expand_spherical_harmonics,
+    expand_ellipsoid_on_image
 )
 from .fit_ellipsoid import (
     expand_points_on_ellipse,
