+++ src/napari_stress/_approximation/expansion_napari.py
@@ -6,12 +6,13 @@ if TYPE_CHECKING:
     import napari
 
 from .._utils import frame_by_frame
-from .expansion import (
-    EllipsoidExpander,
-    EllipsoidImageExpander,
+from .expansion_spherical_harmonics import (
     SphericalHarmonicsExpander,
 )
-
+from .expansion_ellipsoid import (
+    EllipsoidExpander,
+    EllipsoidImageExpander,
+)
 
 @register_function(menu="Points > Fit ellipsoid to pointcloud (n-STRESS)")
 @frame_by_frame
